refactor(issue-533): geo_contracts を基本型trait定義の正本へ統一
