+++ model/geo_core/src/vector_traits.rs
@@ -1,102 +1,6 @@
-//! Vector core traits defined in geo_core.
+//! Compatibility re-exports for Vector traits.
 
-use analysis::abstract_types::Scalar;
-use analysis::linalg::vector::{Vector2, Vector3};
-
-pub trait Vector2DConstructor<T: Scalar> {
-    fn new(x: T, y: T) -> Self;
-    fn zero() -> Self;
-    fn unit_x() -> Self;
-    fn unit_y() -> Self;
-    fn from_tuple(components: (T, T)) -> Self;
-    fn from_analysis_vector(vector: &Vector2<T>) -> Self;
-    fn from_array(components: [T; 2]) -> Self;
-    fn from_polar(magnitude: T, angle: T) -> Self;
-    fn from_vector(other: &Self) -> Self;
-}
-
-pub trait Vector3DConstructor<T: Scalar> {
-    fn new(x: T, y: T, z: T) -> Self;
-    fn zero() -> Self;
-    fn unit_x() -> Self;
-    fn unit_y() -> Self;
-    fn unit_z() -> Self;
-    fn from_tuple(components: (T, T, T)) -> Self;
-    fn from_analysis_vector(vector: &Vector3<T>) -> Self;
-    fn from_array(components: [T; 3]) -> Self;
-    fn from_spherical(magnitude: T, azimuth: T, elevation: T) -> Self;
-    fn from_cylindrical(radial_distance: T, azimuth: T, height: T) -> Self;
-    fn from_vector(other: &Self) -> Self;
-}
-
-pub trait Vector2DProperties<T: Scalar> {
-    fn x(&self) -> T;
-    fn y(&self) -> T;
-    fn components(&self) -> [T; 2];
-    fn to_tuple(&self) -> (T, T);
-    fn to_analysis_vector(&self) -> Vector2<T>;
-    fn length(&self) -> T;
-    fn length_squared(&self) -> T;
-    fn normalize(&self) -> Self;
-    fn try_normalize(&self) -> Option<Self>
-    where
-        Self: Sized;
-}
-
-pub trait Vector3DProperties<T: Scalar> {
-    fn x(&self) -> T;
-    fn y(&self) -> T;
-    fn z(&self) -> T;
-    fn components(&self) -> [T; 3];
-    fn to_tuple(&self) -> (T, T, T);
-    fn to_analysis_vector(&self) -> Vector3<T>;
-    fn length(&self) -> T;
-    fn length_squared(&self) -> T;
-    fn normalize(&self) -> Self;
-    fn try_normalize(&self) -> Option<Self>
-    where
-        Self: Sized;
-}
-
-pub trait Vector2DMeasure<T: Scalar> {
-    fn dot(&self, other: &Self) -> T;
-    fn cross_2d(&self, other: &Self) -> T;
-    fn angle_to(&self, other: &Self) -> Option<T>;
-    fn distance_to(&self, other: &Self) -> T;
-    fn distance_squared_to(&self, other: &Self) -> T;
-    fn magnitude(&self) -> T;
-    fn manhattan_distance(&self) -> T;
-    fn is_parallel_to(&self, other: &Self) -> bool;
-    fn is_perpendicular_to(&self, other: &Self) -> bool;
-    fn project_onto(&self, other: &Self) -> Option<Self>
-    where
-        Self: Sized;
-}
-
-pub trait Vector3DMeasure<T: Scalar> {
-    fn dot(&self, other: &Self) -> T;
-    fn cross_3d(&self, other: &Self) -> Self;
-    fn angle_to(&self, other: &Self) -> Option<T>;
-    fn distance_to(&self, other: &Self) -> T;
-    fn distance_squared_to(&self, other: &Self) -> T;
-    fn magnitude(&self) -> T;
-    fn manhattan_distance(&self) -> T;
-    fn is_parallel_to(&self, other: &Self) -> bool;
-    fn is_perpendicular_to(&self, other: &Self) -> bool;
-    fn project_onto(&self, other: &Self) -> Option<Self>
-    where
-        Self: Sized;
-    fn project_onto_plane(&self, normal: &Self) -> Option<Self>
-    where
-        Self: Sized;
-}
-
-pub trait Vector2DCore<T: Scalar>:
-    Vector2DConstructor<T> + Vector2DProperties<T> + Vector2DMeasure<T>
-{
-}
-
-pub trait Vector3DCore<T: Scalar>:
-    Vector3DConstructor<T> + Vector3DProperties<T> + Vector3DMeasure<T>
-{
-}
+pub use geo_contracts::{
+    Vector2DConstructor, Vector2DCore, Vector2DMeasure, Vector2DProperties, Vector3DConstructor,
+    Vector3DCore, Vector3DMeasure, Vector3DProperties,
+};
